3주차_엔진 아키텍처 및 컴포넌트 시스템 구축
엔진 코어 및 관리자(Manager) 도입
- Engine: SDL 초기화, 메인 루프, 관리자 클래스들의 생명주기를 담당하는 싱글턴 코어 클래스를 구현했습니다.
- SceneManager: FSM(유한 상태 기계) 기반으로 씬의 로딩, 업데이트, 전환을 관리합니다.
- ObjectManager: std::unique_ptr를 사용하여 씬 내 모든 Object의 생명주기를 관리합니다.
- RenderManager: 모든 MeshRenderer를 등록받아 렌더링을 총괄하고, 셰이더/텍스처 등 리소스를 캐싱합니다.
- InputManager: SDL 이벤트를 캡슐화하여 일관된 입력 처리 인터페이스를 제공합니다.

컴포넌트 기반 아키텍처(Component-Based Architecture) 구현
- Object와 Component 기반 클래스를 도입하여, 상속 대신 구성을 통해 객체의 기능을 확장할 수 있는 구조를 마련했습니다.
- MeshRenderer를 구현하여 Object에 렌더링 기능을 부여했습니다.

그래픽스 파이프라인 리팩토링
- Shader: 하드코딩된 셰이더 코드를 외부 파일(.vert, .frag)에서 읽어오도록 개선했습니다.
- Mesh: 정점/인덱스 데이터와 VAO/VBO 생성 로직을 캡슐화하고, CreateCube, CreateSphere 등 절차적 도형 생성 기능을 추가했습니다.

테스트 씬 (MeshesScene) 추가
- 새로 구축된 아키텍처와 시스템이 올바르게 동작하는지 검증하기 위해, 여러 Object와 MeshRenderer 컴포넌트로 구성된 테스트 씬을 작성했습니다.

+++ graphic/source/IndexBuffer.cpp
@@ -2,7 +2,7 @@
 #include <glew.h>
 
 IndexBuffer::IndexBuffer(std::span<const unsigned int> indices)
-	: count(indices.size())
+	: count(static_cast<int>(indices.size()))
 {
 	glCreateBuffers(1, &indicesHandle);
 	glNamedBufferStorage(indicesHandle, indices.size_bytes(), indices.data(), GL_DYNAMIC_STORAGE_BIT);
